requirements.txt
arrow==1.2.3
asgiref==3.5.2
blessed==1.19.1
certifi==2022.9.14
charset-normalizer==2.0.12
Django==4.0
django-cors-headers==3.13.0
django-picklefield==3.1
django-q==1.3.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
idna==3.4
PyJWT==2.5.0
python-dateutil==2.8.2
pytz==2022.2.1
redis==3.5.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.12
wcwidth==0.2.5
